requirements.txt
Flask
gunicorn
opencv-python
numpy
azure-cognitiveservices-vision-computervision
msrest
tenacity
sdnotify
gevent
